Attempt to get check in use case to work

+++ lib/run_pal/use_cases/delete_post.rb
@@ -4,9 +4,11 @@ module RunPal
     def run(inputs)
 
       inputs[:post_id] = inputs[:post_id].to_i
+      inputs[:user_id] = inputs[:user_id].to_i
 
       post = RunPal.db.get_post(inputs[:post_id])
       return failure(:post_does_not_exist) if post.nil?
+      return failure(:user_not_authorized) if inputs[:user_id] != post.creator_id
 
       delete_post(inputs[:post_id])
       deleted = RunPal.db.get_post(inputs[:post_id])
